Fix native crash reports failing to send
raygun4apple requires us to pass in an endpoint. Previously, we would
only pass in the custom endpoint provided by the user, which would
effectively disable native crash reporting unless a custom endpoint was
used.

Now, we pass in the default Raygun endpoint if we have no custom
endpoint available. This allows native exceptions to be reported as
expected.