Fix undefined Promise.prototype.finally
When tracking unhandled Promise rejections, we polyfill the Promise global.
The `promise` polyfill module is split into a number of submodules, that
implement different parts of the Promise specification.

As we were only requiring es6-extensions, Promise.prototype.finally and
Promise.prototype.done were not defined. We can fix this by requiring
the done and finally implementations from the `promise` module.